Update README.md
Hello - request your review and approval on the inclusion of Explore Perl on your list. It would be a good addition under Other Perl List. OR any Resources / Website like section of this article. Would you like to include it in your list?
We'd love to hear your feedback. Thanks!